# openapi.yml
openapi: 3.0.0
info:
  title: Intear Token Indexer API
  version: 1.0.0
  description: Generated by Claude
servers:
  - url: https://prices.intear.tech
...
paths:
...
  /get-token-price:
    get:
      summary: Get price for a specific token
      description: Retrieves the price for a specified token
      tags:
        - Token Prices
      parameters:
        - name: token_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  token_contract_id:
                    type: string
                  price:
                    type: string
              example:
                token_contract_id: "intel.tkn.near"
                price: "0.00000031403698393197025855699805282527104886657659586331948545230257399577125154783353178706988307614885060246835558957366870252978089345781996132921994271391922684056465779502553052217848616272206867258696"
  /price:
    get:
      summary: Get price as a number for a specific token
      description: Retrieves the price as a number for a specified token
      tags:
        - Token Prices
      parameters:
        - name: token_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: number
                format: double
              example: 0.00000031403698393197
  /super-precise-price:
    get:
      summary: Get high-precision price as a string for a specific token
      description: Retrieves the high-precision price as a string for a specified token
      tags:
        - Token Prices
      parameters:
        - name: token_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            text/plain:
              schema:
                type: string
              example: "0.00000031403698393197025855699805282527104886657659586331948545230257399577125154783353178706988307614885060246835558957366870252978089345781996132921994271391922684056465779502553052217848616272206867258696"
...
  /hardcoded/get-token-price:
    get:
      summary: Get hardcoded price for a specific token
      description: Retrieves the hardcoded price for a specified token. Stablecoins may be rounded to 1 if close to 1.
      tags:
        - Token Prices
      parameters:
        - name: token_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  token_contract_id:
                    type: string
                  price:
                    type: string
              example:
                token_contract_id: "intel.tkn.near"
                price: "0.00000031403698393197025855699805282527104886657659586331948545230257399577125154783353178706988307614885060246835558957366870252978089345781996132921994271391922684056465779502553052217848616272206867258696"
  /hardcoded/price:
    get:
      summary: Get hardcoded price as a number for a specific token
      description: Retrieves the hardcoded price as a number for a specified token. Stablecoins may be rounded to 1 if close to 1.
      tags:
        - Token Prices
      parameters:
        - name: token_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: number
                format: double
              example: 0.00000031403698393197
  /hardcoded/super-precise-price:
    get:
      summary: Get hardcoded high-precision price as a string for a specific token
      description: Retrieves the hardcoded high-precision price as a string for a specified token. Stablecoins may be rounded to 1 if close to 1.
      tags:
        - Token Prices
      parameters:
        - name: token_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            text/plain:
              schema:
                type: string
              example: "0.00000031403698393197025855699805282527104886657659586331948545230257399577125154783353178706988307614885060246835558957366870252978089345781996132921994271391922684056465779502553052217848616272206867258696"
...
  /token:
    get:
      summary: Get detailed information for a specific token
      description: Retrieves detailed information for a specific token, including prices, metadata, and supply information
      tags:
        - Token Information
      parameters:
        - name: token_id
          in: query
          required: true
          schema:
            type: string
          description: The identifier of the token to retrieve information for
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenDetail'
              example:
                account_id: "intel.tkn.near"
                price_usd_raw: "0.00000000000000000031403698393197025855699805282527104886657659586331948545230257399577125154783353178706988307614885060246835558957366870252978089345781996132921994271391922684056465779502553052217848616272206867258696"
                price_usd: "0.00000031403698393197025855699805282527104886657659586331948545230257399577125154783353178706988307614885060246835558957366870252978089345781996132921994271391922684056465779502553052217848616272206867258696000000000000"
                price_usd_hardcoded: "0.00000031403698393197025855699805282527104886657659586331948545230257399577125154783353178706988307614885060246835558957366870252978089345781996132921994271391922684056465779502553052217848616272206867258696000000000000"
                main_pool: "REF-4663"
                metadata:
                  name: "INTEAR"
                  symbol: "INTEL"
                  decimals: 18
                total_supply: "100000000000000000000000000000"
                circulating_supply: "100000000000000000000000000000"
                circulating_supply_excluding_team: "98650377280100000000000000001"
                reputation: "Reputable"
                socials:
                  website: "https://intear.tech"
                  telegram: "https://t.me/intearchat"
                  twitter: "https://x.com/intelnear"
                slug: []
                deleted: false
        '404':
          description: Token not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Token not found"
  /token-search:
    get:
      summary: Search for tokens
      description: Search for tokens based on various criteria including name, ticker, slug, contract address, or social handles
      tags:
        - Token Information
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: The search query (full or partial token name, ticker, slug, contract address, social handles, or other information)
        - name: n
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 5
          description: The maximum number of results to return (default is 5)
        - name: rep
          in: query
          required: false
          schema:
            type: string
            enum: [Spam, Unknown, NotFake, Reputable]
          description: The minimum reputation of the results
        - name: acc
          in: query
          required: false
          schema:
            type: string
          description: The account ID of the user. If the user owns or owned a token, it will be slightly boosted in search
        - name: platform
          in: query
          required: false
          schema:
            type: string
          description: The platform (parent account id) to filter tokens by. If the platform is specified, only tokens from that platform will be returned
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TokenDetail'
              example:
                - account_id: "intel.tkn.near"
                  price_usd_raw: "0.00000000000000000031403698393197025855699805282527104886657659586331948545230257399577125154783353178706988307614885060246835558957366870252978089345781996132921994271391922684056465779502553052217848616272206867258696"
                  price_usd: "0.00000031403698393197025855699805282527104886657659586331948545230257399577125154783353178706988307614885060246835558957366870252978089345781996132921994271391922684056465779502553052217848616272206867258696000000000000"
                  price_usd_hardcoded: "0.00000031403698393197025855699805282527104886657659586331948545230257399577125154783353178706988307614885060246835558957366870252978089345781996132921994271391922684056465779502553052217848616272206867258696000000000000"
                  main_pool: "REF-4663"
                  metadata:
                    name: "INTEAR"
                    symbol: "INTEL"
                    decimals: 18
                  total_supply: "100000000000000000000000000000"
                  circulating_supply: "100000000000000000000000000000"
                  circulating_supply_excluding_team: "98650377280100000000000000001"
                  reputation: "Reputable"
                  socials:
                    website: "https://intear.tech"
                    telegram: "https://t.me/intearchat"
                    twitter: "https://x.com/intelnear"
                  slug: []
                  deleted: false
...
  /tokens-advanced:
    get:
      summary: Get advanced token information with market data
      description: Retrieves detailed information for all tokens, including prices, metadata, supply information, liquidity, and volume data
      tags:
        - Token Information
      parameters:
        - name: take
          in: query
          required: false
          schema:
            type: integer
            format: int64
            default: 18446744073709551615  # usize::MAX
          description: Maximum number of tokens to return
        - name: min_reputation
          in: query
          required: false
          schema:
            type: string
            enum: [Spam, Unknown, NotFake, Reputable]
            default: Spam
          description: Filter tokens by minimum reputation level
        - name: account_ids
          in: query
          required: false
          schema:
            type: string
          description: Comma-separated list of account IDs to filter by
        - name: platform
          in: query
          required: false
          schema:
            type: string
          description: Filter tokens by platform (parent account id)
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/TokenDetail'
              example:
                "intel.tkn.near":
                  account_id: "intel.tkn.near"
                  price_usd_raw: "0.00000000000000000031403698393197025855699805282527104886657659586331948545230257399577125154783353178706988307614885060246835558957366870252978089345781996132921994271391922684056465779502553052217848616272206867258696"
                  price_usd: "0.00000031403698393197025855699805282527104886657659586331948545230257399577125154783353178706988307614885060246835558957366870252978089345781996132921994271391922684056465779502553052217848616272206867258696000000000000"
                  price_usd_hardcoded: "0.00000031403698393197025855699805282527104886657659586331948545230257399577125154783353178706988307614885060246835558957366870252978089345781996132921994271391922684056465779502553052217848616272206867258696000000000000"
                  main_pool: "REF-4663"
                  metadata:
                    name: "INTEAR"
                    symbol: "INTEL"
                    decimals: 18
                  total_supply: "100000000000000000000000000000"
                  circulating_supply: "100000000000000000000000000000"
                  circulating_supply_excluding_team: "98650377280100000000000000001"
                  reputation: "Reputable"
                  socials:
                    website: "https://intear.tech"
                    telegram: "https://t.me/intearchat"
                    twitter: "https://x.com/intelnear"
                  slug: []
                  deleted: false
                  liquidity_usd: 1000000.0
                  volume_usd_1h: 50000.0
                  volume_usd_24h: 1200000.0
                  volume_usd_7d: 8400000.0
components:
  schemas:
...
    TokenDetail:
      type: object
      properties:
        account_id:
          type: string
        price_usd_raw:
          type: string
        price_usd:
          type: string
        price_usd_hardcoded:
          type: string
        main_pool:
          type: string
        metadata:
          type: object
          properties:
            name:
              type: string
            symbol:
              type: string
            decimals:
              type: integer
        total_supply:
          type: string
        circulating_supply:
          type: string
        circulating_supply_excluding_team:
          type: string
        reputation:
          type: string
          enum: [Spam, Unknown, NotFake, Reputable]
        socials:
          type: object
          additionalProperties:
            type: string
        slug:
          type: array
          items:
            type: string
        deleted:
          type: boolean
        liquidity_usd:
          type: number
          format: double
          description: Total liquidity in USD across all pools
        volume_usd_1h:
          type: number
          format: double
          description: Trading volume in USD over the last hour
        volume_usd_24h:
          type: number
          format: double
          description: Trading volume in USD over the last 24 hours
        volume_usd_7d:
          type: number
          format: double
          description: Trading volume in USD over the last 7 days
      required:
        - account_id
        - price_usd_raw
        - price_usd
        - price_usd_hardcoded
        - main_pool
        - metadata
        - total_supply
        - circulating_supply
        - circulating_supply_excluding_team
        - reputation
        - socials
        - slug
        - deleted
        - liquidity_usd
        - volume_usd_1h
        - volume_usd_24h
        - volume_usd_7d